refactor: move metadata into input
