+++ packages/create-react-native-library/src/input.ts
@@ -1,3 +1,4 @@
+import { version } from '../package.json';
 import validateNpmPackage from 'validate-npm-package-name';
 import githubUsername from 'github-username';
 import type yargs from 'yargs';
@@ -375,4 +376,35 @@ export async function createQuestions({
     questions: finalQuestions,
     singleChoiceAnswers,
   };
+}
+
+export function createMetadata(answers: Answers) {
+  // Some of the passed args can already be derived from the generated package.json file.
+  const ignoredAnswers: (keyof Answers)[] = [
+    'name',
+    'slug',
+    'description',
+    'authorName',
+    'authorEmail',
+    'authorUrl',
+    'repoUrl',
+    'example',
+    'reactNativeVersion',
+    'local',
+  ];
+
+  type AnswerEntries<T extends keyof Answers = keyof Answers> = [
+    T,
+    Answers[T],
+  ][];
+
+  const libraryMetadata = Object.fromEntries(
+    (Object.entries(answers) as AnswerEntries).filter(
+      ([answer]) => !ignoredAnswers.includes(answer)
+    )
+  );
+
+  libraryMetadata.version = version;
+
+  return libraryMetadata;
 }
